+++ data/src/main/java/com/grusie/data/model/AlarmTurnOffData.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "alarm_turn_off_table")
 data class AlarmTurnOffData(        //알람 끄는 방법 테이블
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     var turnOffWay: String,
     var count: Int,
-    var alarmId: Int
+    var alarmId: Long
 )
